refactor: move speculative output metrics to common.
